requirements.txt
customtkinter
Pillow
cryptography
pyperclip
requests
argon2-cffi